move AppBar component to component package on root level
+ add string
+ remove unused resources
