--- app/src/main/res/menu/menu_diagnose.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Nextcloud Talk - Android Client
-  ~
-  ~ SPDX-FileCopyrightText: 2024 Marcel Hibbe <dev@mhibbe.de>
-  ~ SPDX-License-Identifier: GPL-3.0-or-later
--->
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/copy"
-        android:title="@string/nc_common_copy" />
-    <item
-        android:id="@+id/share"
-        android:title="@string/share" />
-    <item
-        android:id="@+id/send_mail"
-        android:title="@string/send_email" />
-    <item
-        android:id="@+id/create_issue"
-        android:title="@string/create_issue" />
-</menu>
